requirements.txt
PyQt6
PyQt6-WebEngine
google-genai
python-dotenv
aiosqlite
mistune